+++ eesampling/eesampling/model.py
@@ -34,15 +34,15 @@ logger = logging.getLogger(__name__)
 
 
 class StratifiedSampling(object):
-    def __init__(self, train_label="train", test_label="test", output_name="output"):
+    def __init__(self, treatment_label="treatment", pool_label="pool", output_name="output"):
         self.columns = {}
-        self.train_label = train_label
-        self.test_label = test_label
+        self.treatment_label = treatment_label
+        self.pool_label = pool_label
         self.output_name = output_name
         self.trained = False
         self.sampled = False
-        self.data_train = None
-        self.data_test = None
+        self.data_treatment = None
+        self.data_pool = None
         self.data_sample = None
 
     def _chop_outliers(self, df):
@@ -120,30 +120,30 @@ class StratifiedSampling(object):
 
     def fit_and_sample(
         self,
-        df_train,
-        df_test,
+        df_treatment,
+        df_pool,
         n_samples_approx=None,
-        min_n_train_per_bin=0,
+        min_n_treatment_per_bin=0,
         random_seed=1,
-        min_n_sampled_to_n_train_ratio=4,
+        min_n_sampled_to_n_treatment_ratio=4,
         relax_n_samples_approx_constraint=False,
     ):
         """
         Attributes
         ----------
-        df_train: pandas.DataFrame
+        df_treatment: pandas.DataFrame
             dataframe to use for constructing the stratified sampling bins.
-        df_test: pandas.DataFrame
+        df_pool: pandas.DataFrame
             dataframe to sample from according to the constructed stratified sampling bins.
         n_samples_aprox: int
-            approximate number of total samples from df_test. It is approximate because
+            approximate number of total samples from df_pool. It is approximate because
             there may be some slight discrepencies around the total count to ensure
             that each bin has the correct percentage of the total.
-        min_n_train_per_bin: int
-            Minimum number of training samples that must exist in a given bin for 
+        min_n_treatment_per_bin: int
+            Minimum number of treatment samples that must exist in a given bin for 
             it to be considered a non-outlier bin (only applicable if there are 
             cols with fixed_width=True)
-        min_n_sampled_to_n_train_ratio: int
+        min_n_sampled_to_n_treatment_ratio: int
         relax_n_samples_approx_constraint: bool
             If True, treats n_samples_approx as an upper bound, but gets as many comparison group
             meters as available up to n_samples_approx. If False, it raises an exception
@@ -159,26 +159,26 @@ class StratifiedSampling(object):
                 while not completed:
                     logging.info(f"Computing bins: {self.get_all_n_bins_as_str()} ")
                     self.fit(
-                        df_train,
-                        min_n_train_per_bin=min_n_train_per_bin,
+                        df_treatment,
+                        min_n_treatment_per_bin=min_n_treatment_per_bin,
                         random_seed=random_seed,
                     )
                     self.sample(
-                        df_test,
+                        df_pool,
                         n_samples_approx=n_samples_approx,
                         random_seed=random_seed,
                         relax_n_samples_approx_constraint=relax_n_samples_approx_constraint,
                     )
 
                     def _violates_ratio():
-                        n_sampled_to_n_train_ratio = (
-                            self.diagnostics().n_sampled_to_n_train_ratio()
+                        n_sampled_to_n_treatment_ratio = (
+                            self.diagnostics().n_sampled_to_n_treatment_ratio()
                         )
-                        if n_sampled_to_n_train_ratio < min_n_sampled_to_n_train_ratio:
+                        if n_sampled_to_n_treatment_ratio < min_n_sampled_to_n_treatment_ratio:
                             logger.info(
-                                f"Insufficient test data in one of the bins for {col['name']}:"
-                                f"found {n_sampled_to_n_train_ratio}:1 but need "
-                                f"{min_n_sampled_to_n_train_ratio}:1. Using last successful n_bins."
+                                f"Insufficient pool data in one of the bins for {col['name']}:"
+                                f"found {n_sampled_to_n_treatment_ratio}:1 but need "
+                                f"{min_n_sampled_to_n_treatment_ratio}:1. Using last successful n_bins."
                             )
                             return True
                         return False
@@ -191,25 +191,25 @@ class StratifiedSampling(object):
                                 " stratification params are used)."
                             )
                         completed = self.diagnostics().equivalence_passed([col["name"]])
-                        if min_n_sampled_to_n_train_ratio and _violates_ratio():
+                        if min_n_sampled_to_n_treatment_ratio and _violates_ratio():
                             completed = True
                             self.set_n_bins(name, self.get_n_bins(name) - 1)
                         if not completed:
                             self.set_n_bins(name, self.get_n_bins(name) + 1)
                     else:
-                        if min_n_sampled_to_n_train_ratio and _violates_ratio():
+                        if min_n_sampled_to_n_treatment_ratio and _violates_ratio():
                             self.set_n_bins(name, self.get_n_bins(name) - 1)
                             completed = True
                         else:
                             self.set_n_bins(name, self.get_n_bins(name) + 1)
 
         self.fit(
-            df_train, min_n_train_per_bin=min_n_train_per_bin, random_seed=random_seed
+            df_treatment, min_n_treatment_per_bin=min_n_treatment_per_bin, random_seed=random_seed
         )
-        n_treatment = len(df_train)
+        n_treatment = len(df_treatment)
         # if n_samples_approx is None, use the maximum available.
         df_sample = self.sample(
-            df_test, n_samples_approx=n_samples_approx, random_seed=random_seed,
+            df_pool, n_samples_approx=n_samples_approx, random_seed=random_seed,
                         relax_n_samples_approx_constraint=relax_n_samples_approx_constraint,
         )
         self.n_samples_approx = n_samples_approx
@@ -232,29 +232,29 @@ class StratifiedSampling(object):
         col["n_bins"] = n_bins
         self.columns[col_name] = col
 
-    def fit(self, df_train, min_n_train_per_bin=0, random_seed=1):
-        self._check_columns_present(df_train)
-        df_train = self._perturb(self._chop_outliers(df_train), random_seed=random_seed)
-        self.df_train = df_train.copy()
+    def fit(self, df_treatment, min_n_treatment_per_bin=0, random_seed=1):
+        self._check_columns_present(df_treatment)
+        df_treatment = self._perturb(self._chop_outliers(df_treatment), random_seed=random_seed)
+        self.df_treatment = df_treatment.copy()
         self.binning = Binning()
 
-        self.df_train["_outlier_value"] = False
+        self.df_treatment["_outlier_value"] = False
         for name, col in self.columns.items():
 
             if col["min_value_allowed"] is not None:
-                self.df_train.loc[
-                    self.df_train[col["name"]] < col["min_value_allowed"],
+                self.df_treatment.loc[
+                    self.df_treatment[col["name"]] < col["min_value_allowed"],
                     "_outlier_value",
                 ] = True
             if col["max_value_allowed"] is not None:
-                self.df_train.loc[
-                    self.df_train[col["name"]] > col["max_value_allowed"],
+                self.df_treatment.loc[
+                    self.df_treatment[col["name"]] > col["max_value_allowed"],
                     "_outlier_value",
                 ] = True
 
         for name, col in self.columns.items():
             values = (
-                self.df_train.loc[~self.df_train._outlier_value, col["name"]]
+                self.df_treatment.loc[~self.df_treatment._outlier_value, col["name"]]
                 .dropna()
                 .astype(float)
             )
@@ -262,30 +262,30 @@ class StratifiedSampling(object):
                 values, col["name"], col["n_bins"], fixed_width=col["fixed_width"]
             )
 
-        self.data_train = BinnedData(
-            self.df_train, self.binning, min_n_train_per_bin=min_n_train_per_bin
+        self.data_treatment = BinnedData(
+            self.df_treatment, self.binning, min_n_treatment_per_bin=min_n_treatment_per_bin
         )
         self.trained = True
 
     # what kinds of diagnostics?
-    # - explore raw training data
-    # - explore raw test data
-    # - compare training data vs test data, pre-fit
-    # - compare training data vs test data, post-fit
-    # - compare training data vs test data, post-sampled
+    # - explore raw treatment data
+    # - explore raw pool data
+    # - compare treatment data vs pool data, pre-fit
+    # - compare treatment data vs pool data, post-fit
+    # - compare treatment data vs pool data, post-sampled
 
     def diagnostics(self):
         return Diagnostics(model=self)
 
-    def sample(self, df_test, n_samples_approx=None, random_seed=1, relax_n_samples_approx_constraint=False):
-        if not self.trained and data_train is not None:
+    def sample(self, df_pool, n_samples_approx=None, random_seed=1, relax_n_samples_approx_constraint=False):
+        if not self.trained and data_treatment is not None:
             raise ValueError("No model found; please run fit()")
-        self._check_columns_present(df_test)
-        df_test = self._perturb(self._chop_outliers(df_test), random_seed=random_seed)
-        self.data_test = BinnedData(df_test, self.binning)
+        self._check_columns_present(df_pool)
+        df_pool = self._perturb(self._chop_outliers(df_pool), random_seed=random_seed)
+        self.data_pool = BinnedData(df_pool, self.binning)
         df_sample = sample_bins(
-            self.data_train,
-            self.data_test,
+            self.data_treatment,
+            self.data_pool,
             n_samples_approx=n_samples_approx,
             relax_n_samples_approx_constraint=relax_n_samples_approx_constraint,
             random_seed=random_seed,