+++ eesampling/tests/test_diagnostics.py
@@ -27,11 +27,11 @@ from eesampling.model import StratifiedSampling
 
 
 @pytest.fixture
-def diagnostics_obj(df_train, df_test, col_name):
+def diagnostics_obj(df_treatment, df_pool, col_name):
     stratified_sampling_obj = StratifiedSampling()
     stratified_sampling_obj.add_column(col_name, n_bins=4)
     stratified_sampling_obj.fit_and_sample(
-        df_train, df_test, n_samples_approx=len(df_train), random_seed=1
+        df_treatment, df_pool, n_samples_approx=len(df_treatment), random_seed=1
     )
     return stratified_sampling_obj.diagnostics()
 
@@ -41,7 +41,7 @@ def test_equivalence(diagnostics_obj):
     assert equivalence["ks_ok"].all() == True and equivalence["t_ok"].all() == True
 
 
-def test_record_based_equivalence_euclidean(diagnostics_obj, df_train, df_test, df_equiv):
+def test_record_based_equivalence_euclidean(diagnostics_obj, df_treatment, df_pool, df_equiv):
 
     
     equivalence = diagnostics_obj.records_based_equivalence_euclidean(
@@ -50,7 +50,7 @@ def test_record_based_equivalence_euclidean(diagnostics_obj, df_train, df_test, 
     assert equivalence
 
 
-def test_record_based_equivalence_chisquare(diagnostics_obj, df_train, df_test, df_equiv):
+def test_record_based_equivalence_chisquare(diagnostics_obj, df_treatment, df_pool, df_equiv):
 
     
     equivalence = diagnostics_obj.records_based_equivalence_chisquare(